+++ client/src/component/clusterinfo/index.js
@@ -19,9 +19,10 @@ import Actions from './actions';
 import Agents from './agents';
 import Discovery from './discovery';
 import Events from './events';
-import OrchestrateReport from './orchestrate-report';
-import Overview from './overview';
 import Nodes from './nodes';
+import OrchestrateReport from './orchestrate-report';
+import OrchestratorActions from './orchestrator-actions';
+import Overview from './overview';
 
 
 export class OverviewRedirect extends React.Component<{match: any}> {
@@ -68,9 +69,10 @@ export class InnerCluster extends React.Component<Props> {
     const agents = `${root}/agents`;
     const discovery = `${root}/discovery`;
     const events = `${root}/events`;
-    const orchestrate_report = `${root}/orchestrate_report`;
-    const overview = `${root}/overview`;
     const nodes = `${root}/nodes`;
+    const orchestrate_report = `${root}/orchestrate-report`;
+    const orchestrator_actions = `${root}/orchestrators-actions`;
+    const overview = `${root}/overview`;
     return (
       <div>
         <Route exact path={root} component={OverviewRedirect} />
@@ -80,6 +82,7 @@ export class InnerCluster extends React.Component<Props> {
         <Route path={nodes} component={Nodes} />
         <Route path={events} component={Events} />
         <Route path={actions} component={Actions} />
+        <Route path={orchestrator_actions} component={OrchestratorActions} />
         <Route path={orchestrate_report} component={OrchestrateReport} />
       </div>
     );
@@ -94,9 +97,10 @@ export class InnerCluster extends React.Component<Props> {
     const agents = `${root}/agents`;
     const discovery = `${root}/discovery`;
     const events = `${root}/events`;
-    const orchestrate_report = `${root}/orchestrate_report`;
-    const overview = `${root}/overview`;
     const nodes = `${root}/nodes`;
+    const orchestrate_report = `${root}/orchestrate-report`;
+    const orchestrator_actions = `${root}/orchestrators-actions`;
+    const overview = `${root}/overview`;
     return (
       <div className="cluster-tabs">
         <NavLink className="tab" to={overview}>Overview</NavLink>
@@ -105,6 +109,7 @@ export class InnerCluster extends React.Component<Props> {
         <NavLink className="tab" to={nodes}>Nodes</NavLink>
         <NavLink className="tab" to={events}>Events</NavLink>
         <NavLink className="tab" to={actions}>Actions</NavLink>
+        <NavLink className="tab" to={orchestrator_actions}>Orchestrator Actions</NavLink>
         <NavLink className="tab" to={orchestrate_report}>Orchestrate Report</NavLink>
       </div>
     );